Finished work on data manipulation, started working on main model in models.py and the basic training file in run_model.py.
There's been a bit of a situation with the dataset: currently, at each index, only one A matrix is returned for that timestep. However, I think maybe it's supposed to be for the last t timesteps? Not sure... quite tired now so will reread paper in morning.

Also, look into how GCRU is processing X: It seems that there is some fuckery going on that sends the wrong slice of X_list further down the pipeline, however, this may just be due to the above issues with A.

Will leave one more commit tonight with #TODO at the problem lines.